+++ src/lib.rs
@@ -1,1 +1,4 @@
+mod error;
+pub use error::Error;
+
 pub mod network;