+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+<a name="v0.2.0"></a>
+## v0.1.0 (2025-12-29)
+
+#### Features
+
+* Implement `local_ipv4_addresses` which retrieves local IPv4 addresses.
+
+* Implement `local_ipv6_addresses` which retrieves local IPv6 addresses.
+
+* Renamed `get_network_interfaces` to `network_interfaces`.
+
+* Renamed `get_hostname` to `hostname`.
+
 <a name="v0.1.0"></a>
 ## v0.1.0 (2025-12-23)